+++ generators/app/templates/src/graphql/auth/auth.payload.ts
@@ -1,0 +1,13 @@
+
+import { ObjectType, Field } from "type-graphql";
+import { User } from "../../entities/user";
+import { userResponse } from "../../utils/genericTypes";
+
+@ObjectType()
+export class LoginType {
+    @Field()
+    public token: string;
+}
+
+export const LoginPayload = userResponse("LoginPayload", LoginType);
+export const RegisterPayload = userResponse("RegisterPayload", User);