+++ generators/app/templates/src/graphql/project/project.error.ts
@@ -1,0 +1,8 @@
+import { MyError } from "../../utils/error";
+
+export class ProjectSameNameError extends MyError {
+    constructor(projectName: string) {
+        const message = `project ${projectName} already exists`;
+        super(message);
+    }
+}