+++ generators/app/templates/src/graphql/project/project.payload.ts
@@ -1,0 +1,4 @@
+import { userResponse } from "../../utils/genericTypes";
+import { Project } from "../../entities/project";
+
+export const CreateProjectPayload = userResponse("CreateProjectPayload", Project);