+++ generators/app/templates/src/utils/error.ts
@@ -1,0 +1,6 @@
+export class MyError extends Error {
+    public constructor(message?: string) {
+        super(message);
+        Object.setPrototypeOf(this, MyError.prototype);
+    }
+}