+++ generators/app/templates/src/utils/genericTypes.ts
@@ -1,0 +1,18 @@
+import { Field, ObjectType, createUnionType } from "type-graphql";
+
+@ObjectType()
+export class UserError {
+    public constructor(message: string) {
+        this.message = message;
+    }
+
+    @Field()
+    public message: string;
+}
+
+export function userResponse(name: string, objectType: any) {
+    return createUnionType({
+        name,
+        types: () => [objectType, UserError]
+    });
+}