+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
 
+## [2.0.0] - 2019-11-24
+
+### Added
+- Add user error in payload with union fragments
+
+### Changed
+- Standardize arguments
+
+### Removed
+- Lazy relation in entities
+
 ## [1.0.0] - 2019-07-14
 First release